+++ system/src/apps/_module.mjs
@@ -3,4 +3,5 @@ export {default as FalloutPartySleep} from "./FalloutPartySleep.mjs";
 export {default as ItemSelector} from "./CompendiumItemSelectors/ItemSelector.mjs";
 export {default as ItemTypeMenu} from "./ItemTypeMenu.mjs";
 export {default as SalvageJunk} from "./SalvageJunk.mjs";
+export {default as ScavengingTableSettings} from "./ScavengingTableSettings.mjs";
 export {default as SourceFilterSettings} from "./SourceFilterSettings.mjs";