+++ system/src/sheets/_module.mjs
@@ -2,4 +2,5 @@ export {default as FalloutCreatureSheet} from "./FalloutCreatureSheet.mjs";
 export {default as FalloutItemSheet} from "./FalloutItemSheet.mjs";
 export {default as FalloutNpcSheet} from "./FalloutNpcSheet.mjs";
 export {default as FalloutPcSheet} from "./FalloutPcSheet.mjs";
+export {default as FalloutScavengingLocationSheet} from "./FalloutScavengingLocationSheet.mjs";
 export {default as FalloutSettlementSheet} from "./FalloutSettlementSheet.mjs";